feat: Implement IndexController and routes for home, books, about us,…

+++ database/migrations/2026_01_26_134038_my_library_app.php
@@ -38,7 +38,7 @@ return new class extends Migration
             $table->string(column: 'judul');
             $table->string('penulis');
             $table->string('penerbit');
-            $table->integer('rating')->nullable();
+            $table->float('rating')->nullable();
             $table->date('tahun_terbit');
             $table->foreignId('kategori_id')->nullable()->constrained('book_categories')->onDelete('cascade');
             $table->text('deskripsi');
